+++ packages/client/src/index.js
@@ -6,3 +6,6 @@ export {
 export { default as actions } from './actions'
 export { default as reducers } from './reducers'
 export { default as validations } from './validations'
+export { default as withLoader } from './helpers/withLoader'
+export { default as withAuthsome } from './helpers/withAuthsome'
+export { default as withMutationState } from './helpers/withMutationState'