+++ packages/components/packages/Navigation/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.2"></a>
+## [1.0.2](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-navigation@1.0.1...pubsweet-component-navigation@1.0.2) (2018-03-27)
+
+
+### Bug Fixes
+
+* resolve remaining jsx-a11y lint issues ([a75c0de](https://gitlab.coko.foundation/pubsweet/pubsweet/commit/a75c0de))
+
+
+
+
 <a name="1.0.1"></a>
 
 ## [1.0.1](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-navigation@1.0.0...pubsweet-component-navigation@1.0.1) (2018-02-16)